extract sign in dialog logic to hook, allow opening dialog in sign in view
