+++ src/hooks/index.ts
@@ -9,3 +9,4 @@ export * from './useActiveUser'
 export * from './useCheckBrowser'
 export * from './useJoystream'
 export * from './useConnectionStatus'
+export * from './useSignInDialog'
